[enh] WS changes: make sure a source filter is given after 20s, otherwise close the session [enh] WS changes: start to listen changes only after defined a source filter

+++ cesium-plus-pod-core/src/main/java/org/duniter/elasticsearch/rest/node/RestNodeStatsGetAction.java
@@ -72,8 +72,8 @@ public class RestNodeStatsGetAction extends BaseRestHandler {
             // Listeners by source
             if (request.paramAsBoolean("listeners", true)) {
                 mapping.startArray("listeners");
-                Map<String, Integer> sourcesListener = changeService.getUsageStatistics();
-                for (Map.Entry<String, Integer> entry : sourcesListener.entrySet()) {
+                Map<String, Long> sourcesListener = changeService.getUsageStatistics();
+                for (Map.Entry<String, Long> entry : sourcesListener.entrySet()) {
                     mapping.startObject()
                             .field("source", entry.getKey())
                             .field("count", entry.getValue())
